+++ src/card.ts
@@ -341,9 +341,8 @@ export class Card extends AlertUpdatesBase {
         display: flex;
         align-items: center;
         background-repeat: no-repeat;
-        color: var(--spc-card-footer-color, inherit);
-        background-color: var(--spc-card-footer-background-color, var(--spc-player-footer-bg-color, var(--card-background-color, transparent)));
-        background-image: var(--spc-card-footer-background-image, linear-gradient(rgba(0, 0, 0, 0.6), rgb(0, 0, 0)));
+        background-color: var(--spc-footer-background-color, var(--spc-player-footer-bg-color, var(--card-background-color, transparent)));
+        background-image: var(--spc-footer-background-image, linear-gradient(rgba(0, 0, 0, 0.6), rgb(0, 0, 0)));
       }
 
       .spc-card-footer {
@@ -1321,19 +1320,35 @@ export class Card extends AlertUpdatesBase {
       }
     }
 
-    // set footer icon size.
-    if (this.config.footerIconSize) {
-      styleInfo['--spc-footer-icon-size'] = `${this.config.footerIconSize}`;
+    // load basic layout settings.
+    const footerBackgroundColor = this.config.footerBackgroundColor;
+    const footerBackgroundImage = this.config.footerBackgroundImage;
+    const footerIconColor = this.config.footerIconColor;
+    const footerIconColorSelected = this.config.footerIconColorSelected;
+    const footerIconSize = this.config.footerIconSize;
+
+    // set css variables that affect the card footer.
+    if (footerIconColor)
+      styleInfo['--spc-footer-icon-color'] = `${footerIconColor}`;
+    if (footerIconColorSelected)
+      styleInfo['--spc-footer-icon-color-selected'] = `${footerIconColorSelected}`;
+    if (footerIconSize) {
+      styleInfo['--spc-footer-icon-size'] = `${footerIconSize}`;
       styleInfo['--spc-footer-icon-button-size'] = `var(--spc-footer-icon-size, ${FOOTER_ICON_SIZE_DEFAULT}) + 0.75rem`;
     }
-
-    // is player selected, and a footer background color set?
-    // if so, then return vibrant background style;
-    // otherwise, let background color default to the card background color.
-    if ((this.section == Section.PLAYER) && (this.footerBackgroundColor)) {
-      styleInfo['--spc-player-footer-bg-color'] = `${this.footerBackgroundColor || 'transparent'}`;
+    if (footerBackgroundImage)
+      styleInfo['--spc-footer-background-image'] = `${footerBackgroundImage}`;
+    if (footerBackgroundColor) {
+      styleInfo['--spc-footer-background-color'] = `${footerBackgroundColor}`;
     } else {
-      styleInfo['background'] = `unset`;
+      // is player selected, and a footer background color set?
+      // if so, then return vibrant background style;
+      // otherwise, let background color default to the card background color.
+      if ((this.section == Section.PLAYER) && (this.vibrantColorVibrant)) {
+        styleInfo['--spc-player-footer-bg-color'] = `${this.footerBackgroundColor || 'transparent'}`;
+      } else {
+        styleInfo['background'] = `unset`;
+      }
     }
 
     return styleMap(styleInfo);